requirements.txt
alembic==1.11.1
asyncpg==0.27.0
fastapi==0.95.2
fastapi-jwt-auth==0.5.0
mypy==1.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.22
uvicorn==0.22.0
email-validator==1.3.1
passlib==1.7.4
fastapi-mail==1.2.8
Jinja2~=3.1.2
PyJWT~=2.8.0
pydantic~=1.10.7
starlette~=0.27.0
sqlalchemy-stubs~=0.4